refactor home screen
